rofi 1.7.0 fix

+++ rofi/.config/rofi/applets/menu/powermenu.sh
@@ -11,8 +11,8 @@ dir="$HOME/.config/rofi/applets/menu/configs/$style"
 rofi_command="rofi -theme $dir/powermenu.rasi"
 
 uptime=$(uptime -p | sed -e 's/up //g')
-cpu=$(sh ~/.config/rofi/bin/usedcpu)
-memory=$(sh ~/.config/rofi/bin/usedram)
+cpu=$($HOME/.config/rofi/bin/usedcpu)
+memory=$($HOME/.config/rofi/bin/usedram)
 
 # Options
 shutdown=""
@@ -42,57 +42,57 @@ chosen="$(echo -e "$options" | $rofi_command -p "祥  $uptime  |     $cpu  | 
 case $chosen in
     $shutdown)
 		ans=$(confirm_exit &)
-		#if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
+		if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
 			systemctl poweroff
-		#elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
-			#exit 0
-        #else
-			#msg
-        #fi
+		elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
+			exit 0
+        else
+			msg
+        fi
         ;;
     $reboot)
 		ans=$(confirm_exit &)
-		#if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
+		if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
 			systemctl reboot
-		#elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
-			#exit 0
-        #else
-			#msg
-        #fi
+		elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
+			exit 0
+        else
+			msg
+        fi
         ;;
     $lock)
-		#if [[ -f /usr/bin/i3lock ]]; then
-			#i3lock
-		#elif [[ -f /usr/bin/betterlockscreen ]]; then
+		if [[ -f /usr/bin/i3lock ]]; then
+			i3lock
+		elif [[ -f /usr/bin/betterlockscreen ]]; then
 			betterlockscreen -l
-		#fi
+		fi
         ;;
     $suspend)
 		ans=$(confirm_exit &)
-		#if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
+		if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
 			mpc -q pause
 			amixer set Master mute
 			systemctl suspend
-		#elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
-			#exit 0
-        #else
-			#msg
-        #fi
+		elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
+			exit 0
+        else
+			msg
+        fi
         ;;
     $logout)
 		ans=$(confirm_exit &)
-		#if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
-			#if [[ "$DESKTOP_SESSION" == "Openbox" ]]; then
-				#openbox --exit
-			#elif [[ "$DESKTOP_SESSION" == "bspwm" ]]; then
+		if [[ $ans == "yes" || $ans == "YES" || $ans == "y" || $ans == "Y" ]]; then
+			if [[ "$DESKTOP_SESSION" == "Openbox" ]]; then
+				openbox --exit
+			elif [[ "$DESKTOP_SESSION" == "bspwm" ]]; then
 				bspc quit
-			#elif [[ "$DESKTOP_SESSION" == "i3" ]]; then
-				#i3-msg exit
-			#fi
-		#elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
-			#exit 0
-        #else
-			#msg
-        #fi
+			elif [[ "$DESKTOP_SESSION" == "i3" ]]; then
+				i3-msg exit
+			fi
+		elif [[ $ans == "no" || $ans == "NO" || $ans == "n" || $ans == "N" ]]; then
+			exit 0
+        else
+			msg
+        fi
         ;;
 esac
